fix: type int to float

+++ 0x00-python_variable_annotations/5-sum_list.py
@@ -5,7 +5,7 @@ Module 5-sum_list
 from typing import List
 
 
-def sum_list(input_list: List[int]) -> float:
+def sum_list(input_list: List[float]) -> float:
     """
     Takes a list of floats and returns
     their sum
